move connectors to own folder
